+++ changelog.md
@@ -6,6 +6,15 @@ The format is based on [Keep a Changelog](	)
 ### Added
 - Firebase entries for location updating
 
+## - 2017-11-03 - Kent Huang
+### Added
+- Email login
+- Get kits and display kits in tableview
+-Kit add function in user menu
+- Location picker in kit add function
+-Show kit detail when table cell is tapped
+-kit delete function in kit detail view
+
 ## - 2017-11-18 - Yudhvir
 ### Added
 - Center on user location button added